python/requirements.txt
aiohttp
python-telegram-bot
psycopg2-binary
python-dotenv
requests